Refactor character to fragment
