+++ app/src/main/java/com/rafagnin/tvshowcase/presentation/fragment/FavoritesFragment.kt
@@ -1,6 +1,5 @@
 package com.rafagnin.tvshowcase.presentation.fragment
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -8,6 +7,8 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
+import androidx.navigation.fragment.findNavController
+import com.rafagnin.tvshowcase.NavGraphDirections
 import com.rafagnin.tvshowcase.databinding.FragmentFavoritesBinding
 import com.rafagnin.tvshowcase.ext.gone
 import com.rafagnin.tvshowcase.ext.show
@@ -68,9 +69,9 @@ class FavoritesFragment : Fragment(), ShowsAdapter.AdapterCallback {
     }
 
     private fun openDetailScreen(id: Long) {
-//        val intent = Intent(context, ShowDetailFragment::class.java)
-//        intent.putExtra(ShowDetailFragment.ID_EXTRA, id)
-//        startActivity(intent)
+        findNavController().navigate(
+            NavGraphDirections.appToShowdetail(id)
+        )
     }
 
     companion object {
